+++ packages/gatsby-plugin-react-css-modules/src/gatsby-node.js
@@ -1,8 +1,14 @@
 exports.onCreateBabelConfig = ({ actions }, pluginOptions) => {
+  const isDevelopment = process.env.NODE_ENV !== `production`
+
+  pluginOptions.plugins && delete pluginOptions.plugins
+
   actions.setBabelPlugin({
     name: `babel-plugin-react-css-modules`,
     options: {
-      generateScopedName: `[name]--[local]--[hash:base64:5]`,
+      generateScopedName: isDevelopment
+        ? `[name]--[local]--[hash:base64:5]`
+        : `[hash:base64:5]`,
       webpackHotModuleReloading: process.env.NODE_ENV !== `production`,
       ...pluginOptions,
     },