+++ packages/gatsby/src/schema/infer-graphql-input-fields.js
@@ -17,6 +17,7 @@ const {
   getExampleValues,
   extractFieldNames,
   isEmptyObjectOrArray,
+  INVALID_VALUE,
 } = require(`./data-tree-utils`)
 
 const { findLinkedNode } = require(`./infer-graphql-type`)
@@ -264,11 +265,12 @@ export function inferInputObjectStructureFromNodes({
     let key = k
     // Remove fields for traversing through nodes as we want to control
     // setting traversing up not try to automatically infer them.
-    if (isRoot && EXCLUDE_KEYS[key]) return
+    if (value === INVALID_VALUE || (isRoot && EXCLUDE_KEYS[key])) return
 
     if (_.includes(key, `___NODE`)) {
       // TODO: Union the objects in array
-      const nodeToFind = _.isArray(value) ? value[0] : value
+      const isArray = _.isArray(value)
+      const nodeToFind = isArray ? value[0] : value
       const linkedNode = findLinkedNode(nodeToFind)
 
       // Get from cache if found, else store into it
@@ -286,7 +288,7 @@ export function inferInputObjectStructureFromNodes({
         linkedNodeCache[linkedNode.internal.type] = value
       }
 
-      if (_.isArray(value)) {
+      if (isArray) {
         value = [value]
       }
 
@@ -299,7 +301,7 @@ export function inferInputObjectStructureFromNodes({
       prefix: `${prefix}${_.upperFirst(key)}`,
     })
 
-    if (field == null) return
+    if (field === null) return
     inferredFields[createKey(key)] = field
   })