+++ packages/gatsby/src/utils/webpack.config.js
@@ -194,18 +194,6 @@ module.exports = async (
         ])
         break
       case `build-javascript`: {
-        // Minify Javascript only if needed.
-        configPlugins = program.noUglify
-          ? configPlugins
-          : configPlugins.concat([
-              plugins.uglify({
-                uglifyOptions: {
-                  compress: {
-                    drop_console: false,
-                  },
-                },
-              }),
-            ])
         configPlugins = configPlugins.concat([
           plugins.extractText(),
           // Write out stats object mapping named dynamic imports (aka page
@@ -297,6 +285,7 @@ module.exports = async (
     // Common config for every env.
     // prettier-ignore
     let configRules = [
+      rules.mjs(),
       rules.js(),
       rules.yaml(),
       rules.fonts(),
@@ -371,13 +360,7 @@ module.exports = async (
       // modules. But also make it possible to install modules within the src
       // directory if you need to install a specific version of a module for a
       // part of your site.
-      modules: [
-        directoryPath(path.join(`node_modules`)),
-        `node_modules`,
-        // This is head scratching - without it css modules in production will fail
-        // to find module with relative path
-        `./`,
-      ],
+      modules: [directoryPath(path.join(`node_modules`)), `node_modules`],
       alias: {
         gatsby$: directoryPath(path.join(`.cache`, `gatsby-browser-entry.js`)),
         // Using directories for module resolution is mandatory because
@@ -430,7 +413,7 @@ module.exports = async (
     // https://github.com/defunctzombie/package-browser-field-spec); setting
     // the target tells webpack which file to include, ie. browser vs main.
     target: stage === `build-html` || stage === `develop-html` ? `node` : `web`,
-    profile: stage === `production`,
+
     devtool: getDevtool(),
     // Turn off performance hints as we (for now) don't want to show the normal
     // webpack output anywhere.
@@ -455,7 +438,11 @@ module.exports = async (
       splitChunks: {
         name: false,
       },
-      minimize: !program.noUglify,
+      minimizer: [
+        // TODO: maybe this option should be noMinimize?
+        !program.noUglify && plugins.minifyJs(),
+        plugins.minifyCss(),
+      ].filter(Boolean),
     }
   }