+++ www/src/components/layout.js
@@ -1,6 +1,7 @@
 import React from "react"
 import Modal from "react-modal"
 import Helmet from "react-helmet"
+import { SkipNavLink } from "@reach/skip-nav"
 import MdClose from "react-icons/lib/md/close"
 import { navigate, PageRenderer } from "gatsby"
 import presets, { colors } from "../utils/presets"
@@ -19,7 +20,6 @@ import "../fonts/Webfonts/futurapt_demiitalic_macroman/stylesheet.css"
 
 // Other fonts
 import "typeface-spectral"
-import "typeface-space-mono"
 
 let windowWidth
 
@@ -56,12 +56,7 @@ class DefaultLayout extends React.Component {
   }
 
   render() {
-    const {
-      location = {
-        pathname: `/starter-showcase`,
-      },
-    } = this.props // location will be undefined if on 'starter-showcase'
-    const isHomepage = location.pathname === `/`
+    const isHomepage = this.props.location.pathname === `/`
 
     // SEE: template-docs-markdown for why this.props.isSidebarDisabled is here
     const isSidebarDisabled =
@@ -156,17 +151,19 @@ class DefaultLayout extends React.Component {
           <meta name="twitter:site" content="@gatsbyjs" />
           <meta name="og:type" content="website" />
           <meta name="og:site_name" content="GatsbyJS" />
+          <meta name="docsearch:version" content="2.0" />
           <link
             rel="canonical"
             href={`https://gatsbyjs.org${this.props.location.pathname}`}
           />
           <html lang="en" />
         </Helmet>
+        <SkipNavLink css={styles.skipLink}>Skip to main content</SkipNavLink>
         <Banner background={isHomepage ? `#402060` : false}>
-          These are the docs for v2 beta.
+          These are the docs for v2.
           {` `}
           <a
-            href="https://gatsbyjs.org/"
+            href="https://v1.gatsbyjs.org/"
             css={{
               color: `#fff`,
             }}
@@ -211,6 +208,32 @@ class DefaultLayout extends React.Component {
       </div>
     )
   }
+}
+
+const styles = {
+  skipLink: {
+    border: `0`,
+    clip: `rect(0 0 0 0)`,
+    height: 1,
+    width: 1,
+    margin: -1,
+    padding: 0,
+    overflow: `hidden`,
+    position: `absolute`,
+    zIndex: 100,
+    fontSize: `0.85rem`,
+    ":focus": {
+      padding: `0.9rem`,
+      position: `fixed`,
+      top: 10,
+      left: 10,
+      background: `white`,
+      textDecoration: `none`,
+      width: `auto`,
+      height: `auto`,
+      clip: `auto`,
+    },
+  },
 }
 
 export default DefaultLayout