+++ packages/gatsby-image/CHANGELOG.md
@@ -3,6 +3,94 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="2.0.15"></a>
+
+## [2.0.15](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-image/compare/gatsby-image@2.0.14...gatsby-image@2.0.15) (2018-10-15)
+
+**Note:** Version bump only for package gatsby-image
+
+<a name="2.0.14"></a>
+
+## [2.0.14](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-image/compare/gatsby-image@2.0.13...gatsby-image@2.0.14) (2018-10-12)
+
+**Note:** Version bump only for package gatsby-image
+
+<a name="2.0.13"></a>
+
+## [2.0.13](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-image/compare/gatsby-image@2.0.12...gatsby-image@2.0.13) (2018-10-01)
+
+### Bug Fixes
+
+- Address [#8579](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-image/issues/8579) Don't display double alt text. ([#8671](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-image/issues/8671)) ([cb0adee](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-image/commit/cb0adee))
+
+<a name="2.0.12"></a>
+
+## [2.0.12](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-image/compare/gatsby-image@2.0.11...gatsby-image@2.0.12) (2018-09-28)
+
+**Note:** Version bump only for package gatsby-image
+
+<a name="2.0.11"></a>
+
+## [2.0.11](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-image/compare/gatsby-image@2.0.10...gatsby-image@2.0.11) (2018-09-28)
+
+### Bug Fixes
+
+- add missing GatsbyImageProps to TS defs ([#8606](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-image/issues/8606)) ([b6f0d62](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-image/commit/b6f0d62)), closes [/github.com/gatsbyjs/gatsby/blob/76c50aa00653d988043333ceefec7abb34078fca/packages/gatsby-image/src/index.js#L457-L474](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-image/issues/L457-L474)
+
+<a name="2.0.10"></a>
+
+## [2.0.10](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-image/compare/gatsby-image@2.0.9...gatsby-image@2.0.10) (2018-09-26)
+
+**Note:** Version bump only for package gatsby-image
+
+<a name="2.0.9"></a>
+
+## [2.0.9](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-image/compare/gatsby-image@2.0.8...gatsby-image@2.0.9) (2018-09-24)
+
+**Note:** Version bump only for package gatsby-image
+
+<a name="2.0.8"></a>
+
+## [2.0.8](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-image/compare/gatsby-image@2.0.7...gatsby-image@2.0.8) (2018-09-21)
+
+**Note:** Version bump only for package gatsby-image
+
+<a name="2.0.7"></a>
+
+## [2.0.7](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-image/compare/gatsby-image@2.0.6...gatsby-image@2.0.7) (2018-09-20)
+
+**Note:** Version bump only for package gatsby-image
+
+<a name="2.0.6"></a>
+
+## [2.0.6](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-image/compare/gatsby-image@2.0.5...gatsby-image@2.0.6) (2018-09-19)
+
+**Note:** Version bump only for package gatsby-image
+
+<a name="2.0.5"></a>
+
+## [2.0.5](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-image/compare/gatsby-image@2.0.0-rc.4...gatsby-image@2.0.5) (2018-09-17)
+
+**Note:** Version bump only for package gatsby-image
+
+<a name="2.0.0-rc.4"></a>
+
+# [2.0.0-rc.4](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-image/compare/gatsby-image@2.0.0-rc.3...gatsby-image@2.0.0-rc.4) (2018-09-17)
+
+**Note:** Version bump only for package gatsby-image
+
+<a name="2.0.0-rc.3"></a>
+
+# [2.0.0-rc.3](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-image/compare/gatsby-image@2.0.0-rc.2...gatsby-image@2.0.0-rc.3) (2018-09-17)
+
+**Note:** Version bump only for package gatsby-image
+
+<a name="2.0.0-rc.2"></a>
+
+# [2.0.0-rc.2](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-image/compare/gatsby-image@2.0.0-rc.1...gatsby-image@2.0.0-rc.2) (2018-09-12)
+
+**Note:** Version bump only for package gatsby-image
+
 <a name="2.0.0-rc.1"></a>
 
 # [2.0.0-rc.1](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-image/compare/gatsby-image@2.0.0-rc.0...gatsby-image@2.0.0-rc.1) (2018-08-29)