+++ packages/gatsby-transformer-sqip/CHANGELOG.md
@@ -3,6 +3,24 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="2.0.2"></a>
+
+## [2.0.2](https://github.com/gatsbyjs/gatsby/compare/gatsby-transformer-sqip@2.0.1...gatsby-transformer-sqip@2.0.2) (2018-10-16)
+
+**Note:** Version bump only for package gatsby-transformer-sqip
+
+<a name="2.0.1"></a>
+
+## [2.0.1](https://github.com/gatsbyjs/gatsby/compare/gatsby-transformer-sqip@2.0.0...gatsby-transformer-sqip@2.0.1) (2018-10-03)
+
+**Note:** Version bump only for package gatsby-transformer-sqip
+
+<a name="2.0.0"></a>
+
+# [2.0.0](https://github.com/gatsbyjs/gatsby/compare/gatsby-transformer-sqip@2.0.0-rc.7...gatsby-transformer-sqip@2.0.0) (2018-09-17)
+
+**Note:** Version bump only for package gatsby-transformer-sqip
+
 <a name="2.0.0-rc.7"></a>
 
 # [2.0.0-rc.7](https://github.com/gatsbyjs/gatsby/compare/gatsby-transformer-sqip@2.0.0-rc.6...gatsby-transformer-sqip@2.0.0-rc.7) (2018-09-11)